feat: introduce Provider interface for multi-provider runtime abstraction
Add a Provider interface in internal/runtime/ so future container backends
(Podman, Incus, remote Docker) can be added without touching CLI code.
Docker logic moves from Manager into DockerProvider; Manager becomes a thin
facade. NewProvider factory creates its own client, so CLI commands no longer
import the Docker SDK for runtime operations. RuntimeConfig gains a
provider field validated by ValidateRuntime.
